requirements.txt
streamlit>=1.6.0
pandas>=1.3.0
xgboost>=1.5.0
scikit-learn>=0.24.0
joblib>=1.3.0
numpy>=1.21.0
Pillow

